+++ app/mission/bart/+mag/+app/+bart/ResultsManager.m
@@ -1,4 +1,4 @@
-classdef ResultsManager < mag.app.manage.Manager
+classdef ResultsManager < mag.app.manage.ResultsManager
 % RESULTSMANAGER Manager for results of Bartington analysis.
 
     properties (SetAccess = private)
@@ -12,16 +12,17 @@ classdef ResultsManager < mag.app.manage.Manager
             % Create ResultsLayout.
             this.ResultsLayout = uigridlayout(parent);
             this.ResultsLayout.ColumnWidth = "1x";
-            this.ResultsLayout.RowHeight = ["1x", "1x"];
+            this.ResultsLayout.RowHeight = "1x";
 
-            uilabel(this.ResultsLayout, Text = "Not available for Bartington yet.", HorizontalAlignment = "center", VerticalAlignment = "center");
+            % Create science preview.
+            this.instantiateSciencePreview(this.ResultsLayout);
 
             % Reset.
             this.reset();
         end
 
-        function reset(~)
-            % do nothing
+        function reset(this)
+            this.resetSciencePreview();
         end
     end
 
@@ -30,7 +31,12 @@ classdef ResultsManager < mag.app.manage.Manager
         function modelChangedCallback(this, model, ~)
 
             if model.HasAnalysis && model.Analysis.Results.HasScience
-                % do nothing
+
+                results = model.Analysis.Results;
+
+                if results.Input1.HasData && results.Input2.HasData
+                    this.plotSensorPreview(results.Input1.Data, results.Input2.Data, LegendLabels = ["Input1", "Input2"]);
+                end
             else
                 this.reset();
             end